// apps/app/src/modules/builder/types/tableOptions.types.ts
export enum TableOptionsType {
  WITH_HEADER = 'with_header',
  WITH_INSIDE_BORDER = 'with_inside_border',
  WITH_OUTSIDE_BORDER = 'with_outside_border',
  ALIGN_LEFT = 'left',
  ALIGN_CENTER = 'center',
  ALIGN_RIGHT = 'right',
}

export enum TableBorderType {
  NONE = 'none',
  SOLID = 'solid',
  DOTTED = 'dotted',
  DASHED = 'dashed',
  DOUBLE = 'double',
  GROOVE = 'groove',
  RIDGE = 'ridge',
  INSET = 'inset',
  OUTSET = 'outset',
}